Refactor some code to helpers
